+++ taktik/core/social_media/instagram/ui/selectors/support/__init__.py
@@ -2,10 +2,13 @@
 
 from .debug import DebugSelectors, DEBUG_SELECTORS
 from .scroll import ScrollSelectors, SCROLL_SELECTORS
+from .watchdog import WatchdogSelectors, WATCHDOG_SELECTORS
 
 __all__ = [
     "DEBUG_SELECTORS",
     "SCROLL_SELECTORS",
+    "WATCHDOG_SELECTORS",
     "DebugSelectors",
     "ScrollSelectors",
+    "WatchdogSelectors",
 ]